phidata-agent/requirements.txt
# Framework
phidata
openai
# Tools
duckduckgo-search
yfinance
newspaper4k
python-dotenv
lancedb
sqlalchemy